src/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==23.2.1
langchain==0.1.20
langchain-experimental==0.0.60
langchain-openai==0.1.25
PyMuPDF==1.24.3
motor==3.4.0
pydantic-mongo==2.3.0
openai==1.66.3
cohere==5.5.8
qdrant-client==1.10.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2==2.9.10
pgvector==0.4.0
nltk==3.9.1
jsonpointer==2.4
python-dateutil==2.8.2
distro==1.8.0
# Advanced PDF processing with OCR
unstructured[pdf]==0.15.13
pytesseract==0.3.13
pdf2image==1.17.0
pillow==10.4.0
pdfplumber==0.11.4
tabula-py==2.9.3
# Authentication dependencies
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
email-validator==2.1.0
